+++ CHANGELOG.md
@@ -71,6 +71,7 @@ Modified clusters:
 - [matter.js]: Update to 0.13.0-alpha.0-20250423-8917d1d1d.
 - [help]: Updated cli help screen.
 - [logger]: Improved frontend logger cleaning.
+- [parameter]: Added getIntArrayParameter and getStringArrayParameter.
 
 ### Fixed